Add handling errors for commands in container
If any command in docker container returns non-zero value, the RunOverSSH
function returns error. Add handling of this error in Build structure. Also add
preparing commands which return values are not checked for flexibility.
Refactor ShellEvaluator struct. As a side effect of this refactor, the text in
log files is now without weird characters (due to colors in shell).